Move common store test fixtures to conftest.py
